Replace old DefaultDBEntityAnalysisBuilder
