+++ src/main/java/org/dcsa/core/query/DBEntityAnalysis.java
@@ -4,7 +4,6 @@ import org.dcsa.core.extendedrequest.JoinDescriptor;
 import org.dcsa.core.extendedrequest.QueryField;
 import org.dcsa.core.extendedrequest.TableAndJoins;
 import org.dcsa.core.query.impl.DefaultDBEntityAnalysisBuilder;
-import org.dcsa.core.query.impl.NewDBEntityAnalysisBuilder;
 import org.springframework.data.relational.core.sql.Column;
 import org.springframework.data.relational.core.sql.Join;
 import org.springframework.data.relational.core.sql.SqlIdentifier;
@@ -21,7 +20,7 @@ public interface DBEntityAnalysis<T> {
     TableAndJoins getTableAndJoins();
 
     static <T> DBEntityAnalysisBuilder<T> builder(Class<T> entityType) {
-        return new NewDBEntityAnalysisBuilder<>(entityType);
+        return new DefaultDBEntityAnalysisBuilder<>(entityType);
     }
 
     interface DBEntityAnalysisBuilder<T> {
